fill day cards and day screen with api data

+++ app/src/main/java/com/caolancode/weatherapp/data/ForecastDay.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class ForecastDay(
     val date: String,
     @SerializedName("date_epoch")
-    val dateEpoch: Int,
+    val dateEpoch: Long,
     val day: Day,
     val astro: Astro,
     val hour: List<Hour>
